--- lab0/benchen216/helloworld.go
@@ -1,7 +1,0 @@
-package main
-
-import "fmt"
-
-func main() {
-	fmt.Println("hello world")
-}
