--- lab2/input1.txt
@@ -1,2 +1,0 @@
-4
-0
