--- lab2/input2.txt
@@ -1,3 +1,0 @@
-7
-15
-0
