--- lab2/input3.txt
@@ -1,3 +1,0 @@
-113
-10000
-0
