--- lab2/lab2.go
@@ -1,17 +1,0 @@
-package main
-
-import "fmt"
-
-func Sum() {
-
-}
-
-func main() {
-	// Please complete the code to make this program be compiled without error.
-	// Notice that you can only add code in this file.
-	var n int64
-	for {
-		fmt.Scanln(&n)
-
-	}
-}
