--- lab2/output1.txt
@@ -1,1 +1,0 @@
-1+2+3+4=10
