--- lab2/output2.txt
@@ -1,2 +1,0 @@
-1+2+3+4+5+6=21
-1+2+3+4+5+6+8+9+10+11+12+13+15=99
